requirements/dev.txt
-r base.txt
pytest
pytest-django
flake8
black